requirements.txt
asgiref==3.8.1
Brotli==1.1.0
dj-database-url==2.2.0
Django==5.1
djangorestframework==3.15.2
psycopg2-binary==2.9.9
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
whitenoise==6.7.0
